requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
rocketmq-client-python==2.1.1
minio==7.2.0
kserve==0.11.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.4.2
python-json-logger==2.0.7
asyncio==3.4.3
aiohttp==3.8.6
kubernetes==28.1.0 